fix SentResult to able to unmarshal message ID

+++ mailjet_resources.go
@@ -114,6 +114,6 @@ type Attachment struct {
 type SentResult struct {
 	Sent []struct {
 		Email     string
-		MessageID int
+		MessageID int64
 	}
 }
